Fixed flash fallback on Safari
Flash object across browsers behave in different ways. Most notably IE uses ActiveX and require a "classid" and a "movie" param. See http://alistapart.com/article/flashembedcagematch for details.

Having classid as a field seems to break Safari compatibility so I added an 'if IE' conditional.

Haven't compiled the dist files and the html injection is in multiple lines. This is for ease of review; if the author thinks this PR is an appropriate solution then I'll put up a patch with compiled files and the works.

Further recommended changes that I can make if the author approves:
* adding alert that user doesn't have flash/a link to download flash